+++ src/main/java/org/wildfly/extension/cassandra/future/AttributeParser.java
@@ -1,4 +1,4 @@
-package org.wildfly.extension.cassandra;
+package org.wildfly.extension.cassandra.future;
 
 /*
 *
