+++ src/main/java/org/wildfly/extension/cassandra/future/PersistentResourceDefinition.java
@@ -1,4 +1,4 @@
-package org.wildfly.extension.cassandra;
+package org.wildfly.extension.cassandra.future;
 
 import java.util.Collection;
 import java.util.Collections;
